+++ CHANGELOG.md
@@ -118,6 +118,7 @@
 - Fixed a bug where the Asset Indexes utility would display an empty modal once it was finished, if there weren’t any skipped or missing files to report. ([#10992](https://github.com/craftcms/cms/issues/10992))
 - Fixed an error that could occur when saving an address via the `users/save-address` controller action, if the post data included any date attributes. ([#11001](https://github.com/craftcms/cms/issues/11001))
 - Fixed a bug where it wasn’t possible to modify dropdown options within Table fields. ([#11002](https://github.com/craftcms/cms/issues/11002))
+- Fixed a bug where asset transform URLs could contain backslashes on Windows. ([#11004](https://github.com/craftcms/cms/issues/11004))
 
 ## 4.0.0-beta.4 - 2022-04-06